app/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
beam-sdk==0.13.3
blinker==1.4
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.4
cmake==3.26.4
command-not-found==0.3
croniter==1.4.1
cryptography==3.4.8
dataclasses-json==0.5.9
dbus-python==1.2.18
decorator==5.1.1
distro==1.7.0
distro-info===1.1build1
filelock==3.12.2
frozenlist==1.3.3
fsspec==2023.6.0
greenlet==2.0.2
httplib2==0.20.2
huggingface-hub==0.16.4
idna==3.4
importlib-metadata==4.6.4
jeepney==0.7.1
Jinja2==3.1.2
joblib==1.3.1
keyring==23.5.0
langchain==0.0.230
langchainplus-sdk==0.0.20
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lit==16.0.6
MarkupSafe==2.1.3
marshmallow==3.18.0
marshmallow-dataclass==8.5.14
marshmallow-enum==1.5.1
more-itertools==8.10.0
mpmath==1.3.0
multidict==6.0.4
mypy==0.981
mypy-extensions==1.0.0
netifaces==0.11.0
networkx==3.1
nltk==3.8.1
numexpr==2.8.4
numpy==1.25.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.0
openapi-schema-pydantic==1.2.4
packaging==23.1
Pillow==10.0.0
pydantic==1.10.11
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==5.4.1
regex==2023.6.3
requests==2.31.0
safetensors==0.3.1
scikit-learn==1.3.0
scipy==1.11.1
SecretStorage==3.3.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
SQLAlchemy==2.0.18
sympy==1.12
systemd-python==234
tenacity==8.2.2
threadpoolctl==3.1.0
tokenizers==0.13.3
tomli==2.0.1
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
transformers @ git+https://github.com/huggingface/transformers@45025d92f815675e483f32812caa28cce3a960e7
triton==2.0.0
typeguard==2.13.3
typing-inspect==0.9.0
typing_extensions==4.7.1
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.0.3
validators==0.20.0
wadllib==1.3.6
yarl==1.9.2
zipp==1.0.0
